Exchange Selling added

+++ FabrieBank.DAL/Common/DTOs/DTOCurrencyRate.cs
@@ -7,7 +7,7 @@ namespace FabrieBank.DAL.Common.DTOs
         public EnumCurrencyTypes.CurrencyTypes CurrencyCode { get; set; }
         public decimal BanknoteBuyingRate { get; set; }
         public decimal BanknoteSellingRate { get; set; }
-        public double ForexBuyingRate { get; set; }
+        public decimal ForexBuyingRate { get; set; }
         public decimal ForexSellingRate { get; set; }
     }
 }
